refactor: 잔돈 충전 페이지에 이벤트를 추가하는 부분을 MachineManageEvent.js로 분리
