+++ src/components/ProductAddView.js
@@ -1,7 +1,7 @@
 import { HTML_OF_PRODUCT_ADD_PART, HTML_OF_PRODUCT_STATUS_TABLE } from '../utils/html.js';
 import { PRODUCT, VALUES } from '../utils/constants.js';
-import ProductAddCheck from './ProductAddCheck.js';
-import { ERROR_MESSAGE } from '../utils/constants.js';
+// import ProductAddCheck from './ProductAddCheck.js';
+// import { ERROR_MESSAGE } from '../utils/constants.js';
 
 export default class ProductAddView {
   static render() {
